space_app/settings.py
from pydantic import BaseSettings
from dotenv import load_dotenv

load_dotenv()


class Settings(BaseSettings):
    server_host: str = '127.0.0.1'
    server_port: int = 8000
    database_url: str = 'sqlite:///./database.sqlite3'
    jwt_secret: str
    jwt_algorithm: str = 'HS256'
    jwt_expiration: int = 3600

    # class Config:
    #     env_file = ".env"


settings = Settings(
    _env_file='.env',
    _env_file_encoding='utf-8',
)
