requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.89.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.2
orjson==3.8.5
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.4.8
pydantic==1.10.4
pytest==7.2.1
pytest-cov==4.0.0
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
